small stuff, mostly readme

+++ test/OracleReader.t.sol
@@ -10,10 +10,12 @@ contract OracleReaderTest is Test {
     function setUp() public {
         oracle = new OracleReader();
     }
-    uint public constant price_ETH_atBlock_7484876 = 3023816766700000000000;
+
+    uint256 public constant price_ETH_atBlock_7484876 = 3023816766700000000000;
+
     function test_read() public view {
-        uint a;
-        uint b;
+        uint256 a;
+        uint256 b;
         (a, b) = oracle.read();
         assertEq(a, price_ETH_atBlock_7484876);
     }
